\name{fitted.ssdeR}
\alias{fitted.ssdeR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
fitted.ssdeR}
\description{
\code{fitted} method for \code{ssdeR} objects}
\usage{
fitted.ssdeR(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
an object for which the extraction of model fitted values is meaningful.
}
  \item{\dots}{
other arguments.}
}
\details{
extracts the fitted values of \code{ssdeR} objects.
}
\value{
returns vector of fitted values
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Michael Brottrager}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~list }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
